Dynamic frame rates based on latency
More changes to abstract out stuff like FrameTime and MaxAcceleration that can now be variable as the frame rate changes.  It's not quite there yet but it works reasonably well.  The switch between frame rates can still be a little weird so there's probably some physics tweaks I'm missing.

Also, can change the LT in the middle of a game by pausing, changing the number and resuming.