make LT more responsive to latency increases
It's somewhat counterintuitive but it looks like making LT more responsive to a latency spike seems to make the game less jerky on the way up because it gets there quicker.  Also, it results it less chipmunk mode because every second spent in an LT that is too low results in a bit more buffered frames that are played back when it catches up.

Needs to be tested with laggy players.